+++ struct/jumlah_barang.go
@@ -1,5 +1,5 @@
 package _struct
 
 type Jumlah_Barang struct {
-	Jumlah_Barang int `json:"jumlah_barang"`
+	Jumlah_Barang float64 `json:"jumlah_barang"`
 }
