+++ struct/read_max_retur.go
@@ -1,7 +1,7 @@
 package _struct
 
 type Read_max_reture struct {
-	Kode_supplier string `json:"kode_supplier"`
-	Kode_stock    string `json:"kode_stock"`
-	Max_barang    int    `json:"max_barang"`
+	Kode_supplier string  `json:"kode_supplier"`
+	Kode_stock    string  `json:"kode_stock"`
+	Max_barang    float64 `json:"max_barang"`
 }
